Calculate correct content length for UTF-8 string
By using `Content-Length: request.length`, you risk telling the server the wrong byte‐count. Consider this payload:

```
    const response = await client.registerAccount({
      Firstname: "Christöffer",
      Lastname: "Skeppstedt",
      ...
    });
```

When the string contains an "ö" instead of "o", we have:
- Byte count: 930
- String length: 929

When this request is sent to the Trustly API, it responds back with:

```
{
  "name": "JSONRPCError",
  "code": 101,
  "message": "ERROR_MALFORMED_JSON"
}
```

This is likely because the content length header is incorrect. When instead using `Content-Length: Buffer.byteLength(request, 'utf8')`, the request is accepted.